requirements.txt
beautifulsoup4==4.9.0
Rich==10.16.1
selenium==3.141.0
webdriver_manager==3.5.2
emoji==1.6.1
googletrans==3.0.0
google-cloud-dialogflow==2.0.0
nltk==3.6.5
spacy==3.2.1